refactor(helm): jsonnet.go
Moves Jsonnet native func related code into jsonnet.go
